Reworked Screen Shake
Reworked screenshake so it looks better and is more flexible. I replaced the rotational shake with translational, which is how it's usually done. There's two shake modes, sinusodial and noise, of which only noise is used at the moment. Sinusodial has a springier feeling, and could work for hard landings or such.

I also added a bunch of shake for different actions, and made it so that explosions always rock the screen. Exact shake values could use some work though.

For the noise, I ported a javascript library I have used on a couple other of projects. I have got no idea how it performs, but it works just fine for this. Looking at the output though it seems to sometimes generate short, flat plateaus, but it might also just be that noise looks that way. There's still some cleanup to do but that's for the next commit