Kickbomb Reimplementation
Reimplemented the kickbomb.

The kickbomb wraps an explosive, which handles the physics, blowing up and all that while the outer kickbomb adds the extra kicking functionality.

I removed the foot check, as in my opinion the game feels better when you can always return bombs.

If a player is hit by a bomb from the front but are not moving in a direction, the bomb is "picked down" so you it can be returned at your own leisure.