ci: update CI caches. (#690)
This updates each CI workflow to use it's own cache key.
It appears that trying to re-use caches across workflows
doesn't work properly.

This also removes the caching from the release workflow
which will only run for nightly releases and tagged releases
so we don't need to waste the cache space when we can
afford to just wait for the builds.