fix: fix crash related to menu music not looping. (#762)
The music tracks were not set to play on looped mode for menus such as
the character selection or home menu. This would cause the game to crash
when the track had finished and the transition to another menu page
caused the system to try and stop the currently playing song ( which was
no longer playing ).