refactor: editor to use map_interface (#771)
The player interactions handled by the editor.rs file have been
refactored out into a separate struct so that future enhancements to the
editor may use this refactored code instead of copying large sections of
it if similar functionality is used (and will be used as part of my next
update to allow for map randomization).
An example of how this refactoring will help is when large parts of the
map will be updated with one button click. Now simply calling
map_interface.set_tile multiple times is possible instead of
copy-pasting all of the code previously contained in the SetTile
match-case.