fix: implement kill callback for all rendered elements (#777)
All elements that act as a spawner (except for the fish school since the
spawner acts as the parent element holding the inner fish elements) now
have a Spawner component that will be discovered during the "delete
element" process.
I also updated the ElementKillCallback ulid string since I just created
it manually in the previous ticket and only recently learned of a way to
generate a true random ulid string.
We should consider refactoring the fish school spawner to properly
behave like a spawner containing entities.