fix: player spawner kill now removes rendered players (#779)
The player spawners can now be removed and the last player spawner to be
removed will also remove the rendered players. This is done because the
players are able to spawn from a single player spawner (if that is the
only spawner), so removing the rendered players after removing all but
the last spawner would simply cause them to respawn from the existing
player spawner(s).

The ElementKillCallback needed to be brought back into the system in
order to facilitate the complex killing of the rendered player entities.