fix: fixes occasional crash in loading while accessing hats. (#839)
Maybe I should've opened a bug for this (still can if we want one), but
I think I identified the issue.

I've crashed  a handful of times in loading.rs, (GameLoader), here:
```rust
        // Load player hat atlase egui handles
        for hat_handle in &core.player_hats {
            let hat_meta = self.hat_assets.get(&hat_handle.get_bevy_handle()).unwrap();
            let path = hat_meta.atlas.path.clone()
```

I believe this is because we were not waiting for asset to be loaded
(checking if the hat asset had been loading in core_assets_loaded), so
some of the time if not yet loaded by bevy would fail to get asset.