fix(character_movement): Fix sleeping kinematic / has_moved optimization causing AI to sink into ground (#883)
I seem to have set the update of saved positions used to determine if
kinematic has moved in the wrong spot, ended up in a situation in which
AI thinks it had not moved, but was inching down by -0.1 each frame,
sinking into ground.

I was able to repro fairly consistently and feeling pretty confident
this should fix the AI sinking into ground described in #882. (Though I
do not believe this is related to the losing control of player bit).

EDIT: #853 (objects stuck in walls) also appears to be resolved by this
fix.