feat(physics): Add extra colliders to jump_through tiles for collision with dynamics (#931)
define some collision info for jump through wood tiles, used to create
accurate colliders for collision with ragdoll / dynamics.

The standard tile is still there for all intents and purposes. If there
is an extra collider defined, this collider is also added, and is set to
solver group such that dynamics will collider with it.
- This additional collider is not used for collision events.
- if additional collider is present, original tile collider no longer
has solver group for collisions.

This fixes dynamics colliding with the invisible chunk on thin tiles
such as jump through wood.

I think we're done with the trickle of auxiliary features and next PR
should be ragdolls 😆