feat: reimplement blunderbuss (#952)
The new blunderbuss is done, this closes #579.

I used the artwork from #152. `bullet_spread` and `bullet_count` are
open for changes on `assets/elements/item/buss/buss.yaml`. I also made a
quick artwork as the musket bullets explosions are too big for the buss
bullets size, but it's basically a smaller musket explosion. My idea was
a mid-range shotgun but it could be easily modified through `yaml`.
Blunderbuss will spawn in level 1, 3, 4, 7 and 11.



https://github.com/fishfolk/jumpy/assets/107818001/93ee26ef-3e94-49a2-8bce-e7484eb1ed0d