fix: Avoid physics query slowdown over time by rebuilding broadphase every frame (#962)
See #961 for context. There's a bug in parry2d involving incremental
update failing to balance broadphase BVH. Fully rebuilding BVH avoids
this issue, and we can get away with it due to few bodies in our game.
This was our previous behavior before supporting dynamic bodies and
relying on incremental update in physics pipeline, which turns out to be
broken.