feat: Cannon re-implement (#965)
This PR closes #576.

I started this as draft because I think the team feedback is important
before I head to writing the cannon itself.

There are many `.yaml` fields that are used only when holding the
kickbomb, and I wonder if I should delete these from the cannonball (as
it will only work at the lit state) or just leave it there because the
game would just crash if the cannonball is used as non-lit.

I also used `.clone()` while dealing with `SVec<>` and animations in
`kick_bomb.rs` because I think it was the only option, but using
`.clone()` is quite inefficient so I wonder if there's another way of
doing that.