fix(ui): 1-frame glitch when removing player
Gather available input sources after reacting to user inputs so that,
when rendering, we have the most up-to-date list of inputs.

For example, when there is one active slot and the rest are empty, and
the user presses Menu Back, there is a 1 frame glitch where the first
slot says "Press Comma/South" and the rest say "Press
Space/Comma/South". This is because the `player_select_panel` gathered
the available input sources before reacting to user inputs. And when the
user presses Menu Back to remove their player there is now one more
available input, causing the label to be incorrect for that one frame.